pyproject.toml
[tool.poetry]
name = "vinci-train-node"
version = "0.1.0"
description = ""
authors = ["Costa <jorgemdcosta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.10.8"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pytest = "^7.4.2"
pika = "^1.3.2"
redis = "^5.0.0"
websockets = "^11.0.3"
pydantic-settings = "^2.0.3"
python-decouple = "^3.8"
subprocess-run = "^0.0.8"
mlagents = "^0.30.0"
protobuf = "3.20.0"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
